feat(cloudflare): add SvelteKit + Cloudflare Vite plugin
Adds `varlockSvelteKitCloudflarePlugin` exported at
`@varlock/cloudflare-integration/sveltekit` for SvelteKit projects
deploying to Cloudflare Workers via `@sveltejs/adapter-cloudflare`.

The existing `varlockCloudflareVitePlugin` wraps `@cloudflare/vite-plugin`,
which doesn't currently support SvelteKit. The new plugin skips that
wrapper while still injecting a `cloudflare:workers` env loader into the
SvelteKit SSR entry — guarded by a `navigator.userAgent` check and a
dynamic import so SvelteKit's Node-side postbuild steps don't fail
resolving `cloudflare:workers`.

The shared env-loader body (reading `__VARLOCK_ENV` / chunked bindings
from `cloudflare:workers` env and stashing it on
`globalThis.__varlockLoadedEnv`) is extracted to
`shared-ssr-entry-code.ts` and reused by both variants.
