+++ packages/integrations/cloudflare/package.json
@@ -12,7 +12,8 @@
   "types": "dist/index.d.ts",
   "exports": {
     ".": "./dist/index.js",
-    "./init": "./dist/init.js"
+    "./init": "./dist/init.js",
+    "./sveltekit": "./dist/sveltekit.js"
   },
   "bin": {
     "varlock-wrangler": "./bin/varlock-wrangler.js"
