+++ packages/integrations/cloudflare/tsup.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: [
     'src/index.ts',
     'src/init.ts',
+    'src/sveltekit.ts',
     'src/varlock-wrangler.ts',
   ],
